refactor(semantic): merge symbol headers into symbol.h
