+++ src/server/service/feature_router.cpp
@@ -9,8 +9,7 @@
 #include <vector>
 
 #include "command/search_config.h"
-#include "semantic/relation_kind.h"
-#include "semantic/symbol_kind.h"
+#include "semantic/symbol.h"
 #include "server/compiler/compiler.h"
 #include "server/compiler/context_resolver.h"
 #include "server/compiler/indexer.h"
